+++ controller-server/src/main/java/com/yahoo/vespa/hosted/controller/deployment/TestConfigSerializer.java
@@ -1,0 +1,73 @@
+package com.yahoo.vespa.hosted.controller.deployment;
+
+import com.yahoo.config.provision.ApplicationId;
+import com.yahoo.config.provision.ClusterSpec;
+import com.yahoo.config.provision.SystemName;
+import com.yahoo.config.provision.zone.ZoneId;
+import com.yahoo.slime.Cursor;
+import com.yahoo.slime.Slime;
+import com.yahoo.vespa.config.SlimeUtils;
+import com.yahoo.vespa.hosted.controller.api.integration.deployment.JobType;
+
+import java.io.IOException;
+import java.io.UncheckedIOException;
+import java.net.URI;
+import java.util.List;
+import java.util.Map;
+
+/**
+ * Serializes config for integration tests against Vespa deployments.
+ *
+ * @author jonmv
+ */
+public class TestConfigSerializer {
+
+    private final SystemName system;
+
+    public TestConfigSerializer(SystemName system) {
+        this.system = system;
+    }
+
+    /** Returns the config for the tests to run for the given job. */
+    public byte[] testConfig(ApplicationId id,
+                             JobType type,
+                             Map<ZoneId, Map<ClusterSpec.Id, URI>> deployments,
+                             Map<ZoneId, List<String>> clusters) {
+        Slime slime = new Slime();
+        Cursor root = slime.setObject();
+
+        root.setString("application", id.serializedForm());
+        root.setString("zone", type.zone(system).value());
+        root.setString("system", system.value());
+
+        Cursor endpointsObject = root.setObject("endpoints"); // TODO jvenstad: remove.
+        deployments.forEach((zone, endpoints) -> {
+            Cursor endpointArray = endpointsObject.setArray(zone.value());
+            for (URI endpoint : endpoints.values())
+                endpointArray.addString(endpoint.toString());
+        });
+
+        Cursor zoneEndpointsObject = root.setObject("zoneEndpoints");
+        deployments.forEach((zone, endpoints) -> {
+            Cursor clusterEndpointsObject = zoneEndpointsObject.setObject(zone.value());
+            endpoints.forEach((cluster, endpoint) -> {
+                clusterEndpointsObject.setString(cluster.value(), endpoint.toString());
+            });
+        });
+
+        Cursor clustersObject = root.setObject("clusters");
+        clusters.forEach((zone, clusterList) -> {
+            Cursor clusterArray = clustersObject.setArray(zone.value());
+            for (String cluster : clusterList)
+                clusterArray.addString(cluster);
+        });
+
+        try {
+            return SlimeUtils.toJsonBytes(slime);
+        }
+        catch (IOException e) {
+            throw new UncheckedIOException(e);
+        }
+    }
+
+}
