+++ controller-server/src/test/java/com/yahoo/vespa/hosted/controller/deployment/InternalStepRunnerTest.java
@@ -403,20 +403,6 @@ public class InternalStepRunnerTest {
                                            "1554970337.947844\t17491290-v6-1.ostk.bm2.prod.ne1.yahoo.com\t5480\tcontainer\tstderr\twarning\tjava.lang.NullPointerException\\n\\tat org.apache.felix.framework.BundleRevisionImpl.calculateContentPath(BundleRevisionImpl.java:438)\\n\\tat org.apache.felix.framework.BundleRevisionImpl.initializeContentPath(BundleRevisionImpl.java:371)";
 
     @Test
-    public void testConfig() throws IOException {
-        ZoneId zone = ZoneId.from("test", "eu-north-1");
-        byte[] json = InternalStepRunner.testConfig(appId,
-                                                    zone,
-                                                    SystemName.Public,
-                                                    Map.of(zone, Map.of(ClusterSpec.Id.from("ai"),
-                                                                        URI.create("https://server/"))),
-                                                    Map.of(zone, List.of("facts")));
-        byte[] expected = InternalStepRunnerTest.class.getResourceAsStream("/testConfig.json").readAllBytes();
-        assertEquals(new String(SlimeUtils.toJsonBytes(SlimeUtils.jsonToSlime(expected))),
-                     new String(json));
-    }
-
-    @Test
     public void generates_correct_services_xml_test() {
         assertFile("test_runner_services.xml-cd", new String(InternalStepRunner.servicesXml(SystemName.cd, Optional.of("d-2-12-75"))));
     }
